ui: reafctoring
